pyproject.toml
[project]
name = "fpl-web-scrapper"
version = "0.1.0"
description = "Fantasy Premier League player statistics scraper with SQLite storage"
requires-python = ">=3.12"
dependencies = [
    "requests==2.31.0",
    "python-dotenv==1.0.0",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
